+++ src/frontend/devops-repository/src/views/userManage/role.vue
@@ -1,27 +1,31 @@
 <template>
     <div class="role-container" v-bkloading="{ isLoading }">
-        <div class="mb20 flex-align-center">
+        <div class="mt10 flex-between-center">
+            <bk-button class="ml20" icon="plus" theme="primary" @click="createRoleHandler"><span class="mr5">{{ $t('create') }}</span></bk-button>
             <bk-input
-                class="role-search"
                 v-model.trim="role"
-                :right-icon="'bk-icon icon-search'"
+                class="mr20 w250"
+                :placeholder="`共有${roleList.length}个用户组`"
                 clearable
-                :placeholder="`共有${roleList.length}个用户组`">
+                right-icon="bk-icon icon-search">
             </bk-input>
-            <div class="create-user flex-align-center">
-                <bk-button theme="primary" @click.stop="createRoleHandler">{{ $t('create') + $t('userGroup') }}</bk-button>
-            </div>
         </div>
         <bk-table
-            class="role-table"
-            height="calc(100% - 80px)"
+            class="mt10 role-table"
+            height="calc(100% - 62px)"
             :data="filterRoleList"
             :outer-border="false"
             :row-border="false"
-            :row-style="{ cursor: 'pointer' }"
             size="small"
-            @row-click="showUsers"
-        >
+            @row-click="showUsers">
+            <template #empty>
+                <empty-data :is-loading="isLoading" :search="Boolean(role)">
+                    <template v-if="!Boolean(role)">
+                        <span class="ml10">暂无用户组数据，</span>
+                        <bk-button text @click="createRoleHandler">即刻创建</bk-button>
+                    </template>
+                </empty-data>
+            </template>
             <bk-table-column label="用户组名称" prop="name" width="200"></bk-table-column>
             <bk-table-column label="描述" prop="description"></bk-table-column>
             <bk-table-column :label="$t('operation')" width="150">
@@ -31,56 +35,48 @@
                 </div>
             </bk-table-column>
         </bk-table>
-        <bk-dialog
+        <canway-dialog
             v-model="editRoleConfig.show"
             theme="primary"
             width="500"
-            :close-icon="false"
-            header-position="center"
-            :title="editRoleConfig.id ? '编辑用户组' : '添加用户组'">
+            :title="editRoleConfig.id ? '编辑用户组' : '创建用户组'"
+            @cancel="editRoleConfig.show = false">
             <bk-form :label-width="80" :model="editRoleConfig" :rules="rules" ref="roleForm">
                 <bk-form-item label="名称" :required="true" property="name" error-display-type="normal">
-                    <bk-input v-model.trim="editRoleConfig.name"></bk-input>
+                    <bk-input v-model.trim="editRoleConfig.name" maxlength="32" show-word-limit></bk-input>
                 </bk-form-item>
                 <bk-form-item label="简介">
-                    <bk-input type="textarea" v-model.trim="editRoleConfig.description"></bk-input>
+                    <bk-input type="textarea" v-model.trim="editRoleConfig.description" maxlength="200"></bk-input>
                 </bk-form-item>
             </bk-form>
             <div slot="footer">
-                <bk-button theme="primary" @click="confirm">确定</bk-button>
-                <bk-button @click="editRoleConfig.show = false">取消</bk-button>
+                <bk-button @click="editRoleConfig.show = false">{{ $t('cancel') }}</bk-button>
+                <bk-button class="ml10" theme="primary" @click="confirm">{{ $t('confirm') }}</bk-button>
             </div>
-        </bk-dialog>
+        </canway-dialog>
         <bk-sideslider
             class="show-userlist-sideslider"
             :quick-close="true"
             :is-show.sync="editRoleUsers.show"
             :title="editRoleUsers.title"
             :width="500">
-            <div class="show-userlist-content" slot="content">
-                <div class="add-user flex-align-center">
-                    <div style="width:280px">
-                        <bk-select class="select-user"
-                            v-model="editRoleUsers.addUsers"
-                            multiple
-                            searchable
-                            placeholder="请选择用户"
-                            :enable-virtual-scroll="Object.values(userList).length > 3000"
-                            :list="Object.values(userList).filter(user => user.id !== 'anonymous')">
-                            <bk-option v-for="option in Object.values(userList).filter(user => user.id !== 'anonymous')"
-                                :key="option.id"
-                                :id="option.id"
-                                :name="option.name">
-                            </bk-option>
-                        </bk-select>
-                    </div>
-                    <bk-button theme="primary" class="ml10" @click="handleAddUsers">添加</bk-button>
-                    <bk-button theme="warning" class="ml10" @click="handleDeleteUsers">批量移除</bk-button>
+            <template #content>
+                <div class="m10 flex-align-center">
+                    <bk-tag-input
+                        class="w250"
+                        v-model="editRoleUsers.addUsers"
+                        :list="Object.values(selectList)"
+                        :search-key="['id', 'name']"
+                        placeholder="添加用户，按Enter键确认"
+                        trigger="focus"
+                        allow-create>
+                    </bk-tag-input>
+                    <bk-button v-if="editRoleUsers.addUsers.length" theme="primary" class="ml10" @click="handleAddUsers">添加</bk-button>
+                    <bk-button v-if="editRoleUsers.deleteUsers.length" theme="warning" class="ml10" @click="handleDeleteUsers">批量移除</bk-button>
                 </div>
                 <bk-table
-                    class="mt20"
                     :data="editRoleUsers.users"
-                    :max-height="1055"
+                    height="calc(100% - 62px)"
                     stripe
                     border
                     size="small"
@@ -97,7 +93,7 @@
                         </template>
                     </bk-table-column>
                 </bk-table>
-            </div>
+            </template>
         </bk-sideslider>
     </div>
 </template>
@@ -139,28 +135,28 @@
         },
         computed: {
             ...mapState(['userList']),
-            projectId () {
-                return this.$route.params.projectId
-            },
             filterRoleList () {
                 return this.roleList.filter(v => v.name.indexOf(this.role) !== -1)
+            },
+            selectList () {
+                return Object.values(this.userList)
+                    .filter(v => v.id !== 'anonymous')
+                    .filter(v => !~this.editRoleUsers.users.findIndex(w => w === v.id))
             }
         },
         created () {
-            this.getRoleList()
+            this.getRoleListHandler()
         },
         methods: {
             ...mapActions([
-                'getRepoRoleList',
+                'getRoleList',
                 'createRole',
                 'editRole',
                 'deleteRole'
             ]),
-            getRoleList () {
+            getRoleListHandler () {
                 this.isLoading = true
-                return this.getRepoRoleList({
-                    projectId: this.projectId
-                }).then(res => {
+                return this.getRoleList().then(res => {
                     this.roleList = res
                 }).finally(() => {
                     this.isLoading = false
@@ -178,9 +174,11 @@
                 }
             },
             handleAddUsers () {
+                if (!this.editRoleUsers.addUsers.length) return
                 this.editRoleMixin([].concat(this.editRoleUsers.users, this.editRoleUsers.addUsers), '新增用户成功')
             },
             handleDeleteUsers () {
+                if (!this.editRoleUsers.deleteUsers.length) return
                 this.editRoleMixin(this.editRoleUsers.users.filter(v => !this.editRoleUsers.deleteUsers.find(w => w === v)), '移除用户成功')
             },
             editRoleMixin (userIds, message) {
@@ -188,7 +186,6 @@
                 return this.editRole({
                     id: this.editRoleUsers.id,
                     body: {
-                        projectId: this.projectId,
                         userIds
                     }
                 }).then(() => {
@@ -197,11 +194,8 @@
                         message
 
                     })
-                    this.getRepoRoleList({
-                        projectId: this.projectId
-                    }).then(res => {
-                        this.roleList = res
-                        this.showUsers(res.find(v => v.id === this.editRoleUsers.id))
+                    this.getRoleListHandler().then(() => {
+                        this.showUsers(this.roleList.find(v => v.id === this.editRoleUsers.id))
                     })
                 }).finally(() => {
                     this.editRoleUsers.loading = false
@@ -236,9 +230,9 @@
                 fn({
                     id: this.editRoleConfig.id,
                     body: {
-                        projectId: this.projectId,
                         name,
-                        description
+                        description,
+                        type: 'SYSTEM'
                     }
                 }).then(res => {
                     this.$bkMessage({
@@ -246,26 +240,25 @@
                         message: (this.editRoleConfig.id ? '编辑用户组' : '新建用户组') + this.$t('success')
                     })
                     this.editRoleConfig.show = false
-                    this.getRoleList()
+                    this.getRoleListHandler()
                 }).finally(() => {
                     this.editRoleConfig.loading = false
                 })
             },
             deleteRoleHandler ({ id, name }) {
-                this.$bkInfo({
-                    type: 'error',
-                    title: this.$t('deleteRoleTitle', [name]),
-                    subTitle: this.$t('deleteRoleSubTitle'),
-                    showFooter: true,
+                this.$confirm({
+                    theme: 'danger',
+                    message: this.$t('deleteRoleTitle', [name]),
+                    subMessage: this.$t('deleteRoleSubTitle'),
                     confirmFn: () => {
-                        this.deleteRole({
+                        return this.deleteRole({
                             id
                         }).then(() => {
                             this.$bkMessage({
                                 theme: 'success',
                                 message: this.$t('delete') + this.$t('success')
                             })
-                            this.getRoleList()
+                            this.getRoleListHandler()
                         })
                     }
                 })
@@ -275,8 +268,8 @@
 </script>
 <style lang="scss" scoped>
 .role-container {
-    height: calc(100% + 40px);
-    margin-bottom: -40px;
+    height: 100%;
+    overflow: hidden;
     .role-search {
         width: 250px;
     }
@@ -310,18 +303,5 @@
     }
 }
 .show-userlist-sideslider{
-    .bk-dialog-header-inner {
-        font-size: 16px;
-        text-align: center;
-    }
-    .show-userlist-content{
-        height: 100%;
-        min-height: 400px;
-        padding: 20px;
-    }
-    .show-resource-content{
-        height: 100%;
-        padding: 20px;
-    }
 }
 </style>
